cleanup organization of jsonrpc tests
