+++ tests/trinity/core/chain-management/test_initialize_data_dir.py
