+++ tests/trinity/core/chain-management/test_is_data_dir_initialized.py
