+++ tests/trinity/core/chain-management/test_is_database_initialized.py
