+++ tests/trinity/core/chains-utils/test_chain_config_object.py
