+++ tests/trinity/core/cli/test_trinity_repl.py
