+++ tox.ini
@@ -18,17 +18,17 @@ passenv =
     PYTEST_ADDOPTS
     TRAVIS_EVENT_TYPE
 commands=
-    core: py.test {posargs:tests/core/ tests/rpc/ -k 'not test_rpc_against_fixtures'}
+    core: py.test {posargs:tests/core/}
     trinity: py.test {posargs:tests/trinity/}
     p2p: py.test {posargs:tests/p2p}
     database: py.test {posargs:tests/database}
-    rpc-blockchain: py.test {posargs:tests/rpc/test_rpc_fixtures.py -k 'not GeneralStateTests'}
-    rpc-state-frontier: py.test {posargs:tests/rpc/test_rpc_fixtures.py -k 'GeneralStateTests and not stQuadraticComplexityTest and Frontier'}
-    rpc-state-homestead: py.test {posargs:tests/rpc/test_rpc_fixtures.py -k 'GeneralStateTests and not stQuadraticComplexityTest and Homestead'}
-    rpc-state-eip150: py.test {posargs:tests/rpc/test_rpc_fixtures.py -k 'GeneralStateTests and not stQuadraticComplexityTest and EIP150'}
-    rpc-state-eip158: py.test {posargs:tests/rpc/test_rpc_fixtures.py -k 'GeneralStateTests and not stQuadraticComplexityTest and EIP158'}
-    rpc-state-byzantium: py.test {posargs:tests/rpc/test_rpc_fixtures.py -k 'GeneralStateTests and not stQuadraticComplexityTest and Byzantium'}
-    rpc-state-quadratic: py.test {posargs:tests/rpc/test_rpc_fixtures.py -k 'GeneralStateTests and stQuadraticComplexityTest'}
+    rpc-blockchain: py.test {posargs:tests/trinity/json-fixtures-over-rpc/test_rpc_fixtures.py -k 'not GeneralStateTests'}
+    rpc-state-frontier: py.test {posargs:tests/trinity/json-fixtures-over-rpc/test_rpc_fixtures.py -k 'GeneralStateTests and not stQuadraticComplexityTest and Frontier'}
+    rpc-state-homestead: py.test {posargs:tests/trinity/json-fixtures-over-rpc/test_rpc_fixtures.py -k 'GeneralStateTests and not stQuadraticComplexityTest and Homestead'}
+    rpc-state-eip150: py.test {posargs:tests/trinity/json-fixtures-over-rpc/test_rpc_fixtures.py -k 'GeneralStateTests and not stQuadraticComplexityTest and EIP150'}
+    rpc-state-eip158: py.test {posargs:tests/trinity/json-fixtures-over-rpc/test_rpc_fixtures.py -k 'GeneralStateTests and not stQuadraticComplexityTest and EIP158'}
+    rpc-state-byzantium: py.test {posargs:tests/trinity/json-fixtures-over-rpc/test_rpc_fixtures.py -k 'GeneralStateTests and not stQuadraticComplexityTest and Byzantium'}
+    rpc-state-quadratic: py.test {posargs:tests/trinity/json-fixtures-over-rpc/test_rpc_fixtures.py -k 'GeneralStateTests and stQuadraticComplexityTest'}
     transactions: py.test {posargs:tests/json-fixtures/test_transactions.py}
     vm: py.test {posargs:tests/json-fixtures/test_virtual_machine.py}
     native-blockchain: py.test {posargs:tests/json-fixtures/test_blockchain.py}
